+++ packages/respect-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @redocly/respect-core
 
+## 2.28.0
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.28.0.
+
 ## 2.27.1
 
 ### Patch Changes